Fix nits.
Adjust docstrings, imports, indentations, refactorings.
